+++ src/app/grouping/[groupId]/board/[boardId]/components/BoardHeader.client.tsx
@@ -1,23 +1,25 @@
 'use client';
 
-import { Header } from '@/components/NavigationBar';
+import { Header } from '@/components/Header';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
-export default function ProfilePraiseHeader() {
+export default function BoardHeader() {
   const router = useRouter();
+
   return (
     <Header
-      text="받은 칭찬"
       leftNode={
         <Image
           alt="back"
           src="/assets/arrow_back.svg"
-          width={10}
-          height={10}
+          width={8}
+          height={30}
           onClick={() => router.back()}
+          className="cursor-pointer"
         />
       }
+      text="게시판"
     />
   );
 }
