+++ src/app/grouping/[groupId]/feedback/components/FeedbackHeader.tsx
@@ -1,14 +1,18 @@
 'use client';
-
-import { Header } from '@/components/NavigationBar';
+import { Header } from '@/components/Header';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
-export default function BoardHeader() {
+interface FeedbackHeaderProps {
+  text: string;
+}
+
+export default function FeedbackHeader({ text }: FeedbackHeaderProps) {
   const router = useRouter();
 
   return (
     <Header
+      text={text}
       leftNode={
         <Image
           alt="back"
@@ -19,7 +23,6 @@ export default function BoardHeader() {
           className="cursor-pointer"
         />
       }
-      text="게시판"
     />
   );
 }
