+++ src/app/grouping/[groupId]/feedback/components/praise/PraiseComponent.tsx
@@ -1,7 +1,7 @@
 'use client';
 import PraiseCardList from './PraiseCardList';
 import { BottomFixedButton } from '@/components/common/Button';
-import { Header } from '@/components/NavigationBar';
+import { Header } from '@/components/Header';
 import { Spacing } from '@/components/common/Spacing';
 import Image from 'next/image';
