+++ src/app/grouping/create/components/CreateHeader.tsx
@@ -1,23 +1,24 @@
 'use client';
-
-import { Header } from '@/components/NavigationBar';
+import { Header } from '@/components/Header';
 import Image from 'next/image';
 import { useRouter } from 'next/navigation';
 
-export default function ProfileMatesHeader() {
+export default function CreateHeader() {
   const router = useRouter();
   return (
     <Header
-      text="모임 후기"
       leftNode={
         <Image
           alt="back"
           src="/assets/arrow_back.svg"
-          width={10}
-          height={10}
-          onClick={() => router.back()}
+          width={8}
+          height={30}
+          onClick={() => {
+            router.back();
+          }}
         />
       }
+      text="모임 개설하기"
     />
   );
 }
