+++ src/app/join/components/JoinHeader.server.tsx
@@ -1,5 +1,5 @@
 'use client';
-import Header from '@/components/Header/Header';
+import { Header } from '@/components/Header';
 import Image from 'next/image';
 
 interface JoinHeaderProps {
